requirements.txt
flask
mysql-connector-python
flask-wtf
email-validator
